Exception Handling

+++ src/main/java/com/leonardo/dscommerce/services/ProductService.java
@@ -3,14 +3,16 @@ package com.leonardo.dscommerce.services;
 import com.leonardo.dscommerce.DTO.ProductDTO;
 import com.leonardo.dscommerce.entities.Product;
 import com.leonardo.dscommerce.repositories.ProductRepository;
+import com.leonardo.dscommerce.services.exceptions.DatabaseException;
+import com.leonardo.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -20,9 +22,10 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-        ProductDTO dto = new ProductDTO(product);
+        Product result = repository.findById(id)
+                .orElseThrow(
+                        () -> new ResourceNotFoundException("Resource Not Found"));
+        ProductDTO dto = new ProductDTO(result);
         return dto;
     }
 
@@ -42,16 +45,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
-
+        try{
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }catch(EntityNotFoundException e){
+            throw new ResourceNotFoundException("Resource Not Found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource Not Found");
+        }
+        try{
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Fail of Referential Integrity");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
